+++ pages/admin/calendar.tsx
@@ -45,7 +45,7 @@ import EmailModal from "../../views/admin/emailModal";
 import TextareaField from "../../components/formFields/textareaField";
 import { useUser, useSupabaseClient } from "@supabase/auth-helpers-react";
 import { createClient } from "contentful-management";
-import ShowArtworkModal from "../../views/admin/showArtworkModal";
+import AdditionalMenu from "../../views/admin/additionalMenu";
 
 const baseUrl = process.env.NEXT_PUBLIC_SITE_URL
   ? `${process.env.NEXT_PUBLIC_SITE_URL}`
@@ -571,7 +571,7 @@ function Calendar() {
               <TfiReload size={20} />
             )}
           </button>
-          <ShowArtworkModal />
+          <AdditionalMenu />
         </div>
         {/* <DropdownMenu.Root
           open={addDropdownOpen}
